(Not working) add automotive demo phase 0 proof-of-concept 2
This proof-of-concept explores the idea that the error-prone program might unknowningly overwrite some memory (rather than crashing and the compartment unwinding). In our case, due to a bounds error ("<=" instead of "<"), this memory is the acceleration of the car, casing the car to rapidly accelerate!

If running on a legacy (non-CHERI) system, this should cause the effects described above to occur. If running on a CHERI system, this out-of-bounds access should be caught as a result of the appropriate capabilties, and the write will not occur, keeping the car safe despite the error.

Unfortunately this proof of concept does not work - the capability exists for the entire struct allocation and so the contiguous memory allocation does not cause an exception in CHERI. We cannot easily just make two separate allocations because there is no guarantee they are contiguous :(

I'm just keeping this for posterity for now