Other changes to bump system clock frequency
Other changes that are necessary (outside of bumping `cheriot-rtos` to
bump the system clock frequency for all demos. This includes updating
the relevant compile flag, and also bumping the revision of the
`sonata-system` submodule (used for building the legacy component of the
automotive demo) to the latest main, which now uses a frequency of 40
MHz instead of 30 MHz.